requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
chromadb
tavily-python